+++ stac_fastapi/api/stac_fastapi/api/__init__.py
@@ -1,1 +1,1 @@
-"""api submodule."""
+"""Api submodule."""