+++ stac_fastapi/api/stac_fastapi/api/config.py
@@ -3,11 +3,12 @@ import enum
 
 
 # TODO: Move to stac-pydantic
-# Does that make sense now? The shift to json schema rather than a well-known enumeration makes that less obvious.
+# Does that make sense now? The shift to json schema rather than a well-known
+# enumeration makes that less obvious.
 class ApiExtensions(enum.Enum):
     """Enumeration of available stac api extensions.
 
-    Ref: https://github.com/radiantearth/stac-api-spec/tree/master/extensions
+    Ref: https://github.com/stac-api-extensions
     """
 
     context = "context"