+++ stac_fastapi/sqlalchemy/stac_fastapi/sqlalchemy/models/__init__.py
@@ -1,1 +1,0 @@
-"""stac_fastapi.postgres.models module."""